api/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
numpy==1.19.5
pickleshare==0.7.5
gcsfs==2021.11.1
gunicorn
pandas==1.1.4
arcgis==1.9.0
pyarrow==6.0.1
google-cloud-secret-manager>=2.7.2
google.cloud.storage
google.cloud.bigquery